requirements.txt
flask
psycopg2-binary
requests
werkzeug
tqdm
urllib3